+++ packages/intents-sdk/src/bridges/hot-bridge/hot-bridge.ts
@@ -23,6 +23,7 @@ import type {
 	FeeEstimation,
 	NearTxInfo,
 	ParsedAssetInfo,
+	QuoteOptions,
 	RouteConfig,
 	TxInfo,
 	TxNoInfo,
@@ -44,6 +45,7 @@ import {
 import { parseDefuseAssetId } from "../../lib/parse-defuse-asset-id";
 import { getFeeQuote } from "../../lib/estimate-fee";
 import { validateAddress } from "../../lib/validateAddress";
+import isHex from "../../lib/hex";
 
 export class HotBridge implements Bridge {
 	protected env: NearIntentsEnv;
@@ -238,7 +240,7 @@ export class HotBridge implements Bridge {
 
 	async estimateWithdrawalFee(args: {
 		withdrawalParams: Pick<WithdrawalParams, "assetId" | "destinationAddress">;
-		quoteOptions?: { waitMs: number };
+		quoteOptions?: QuoteOptions;
 		logger?: ILogger;
 	}): Promise<FeeEstimation> {
 		const assetInfo = this.parseAssetId(args.withdrawalParams.assetId);
@@ -278,6 +280,7 @@ export class HotBridge implements Bridge {
 		routeConfig: RouteConfig;
 		signal?: AbortSignal;
 		retryOptions?: RetryOptions;
+		logger?: ILogger;
 	}): Promise<TxInfo | TxNoInfo> {
 		const nonces = await this.hotSdk.near.parseWithdrawalNonces(
 			args.tx.hash,
@@ -306,6 +309,14 @@ export class HotBridge implements Bridge {
 					return { hash: null };
 				}
 				if (typeof status === "string") {
+					// HOT returns string hexified raw bytes without 0x prefix, any other value should be ignored.
+					if (!isHex(status)) {
+						args.logger?.warn(
+							"HOT Bridge incorrect destination tx hash detected",
+							{ value: status },
+						);
+						return { hash: null };
+					}
 					return {
 						hash:
 							"chain" in args.routeConfig &&