+++ packages/intents-sdk/src/intents/interfaces/salt-manager.ts
@@ -1,0 +1,6 @@
+import type { Salt } from "../expirable-nonce";
+
+export interface ISaltManager {
+	getCachedSalt(): Promise<Salt>;
+	refresh(): Promise<Salt>;
+}