+++ packages/contract-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @defuse-protocol/contract-types
 
+## 0.2.0
+
+### Minor Changes
+
+- 9dfbecf: Update ABI of the contract.
+
 ## 0.1.2
 
 ### Patch Changes